fix bug

+++ database/migrations/2020_05_11_081856_create_user_points_table.php
@@ -19,7 +19,7 @@ class CreateUserPointsTable extends Migration
             $table->foreignId('user_id');
             $table->foreign("user_id")->references("id")->on("users");
 
-            $table->decimal("amount")->default(0);
+            $table->integer("amount")->default(0);
             $table->string("notes")->collation("utf8_unicode_ci")->nullable(true);
             $table->timestamps();
         });
